Address PR feedback * Move ncproxy related package to unified location * Add manifest to ncproxy binary
