+++ cmd/ncproxy/ncproxy.go
@@ -9,8 +9,8 @@ import (
 	"github.com/Microsoft/hcsshim/hcn"
 	"github.com/Microsoft/hcsshim/internal/computeagent"
 	"github.com/Microsoft/hcsshim/internal/log"
-	"github.com/Microsoft/hcsshim/internal/ncproxynetworking"
-	"github.com/Microsoft/hcsshim/internal/ncproxystore"
+	ncproxynetworking "github.com/Microsoft/hcsshim/internal/ncproxy/networking"
+	ncproxystore "github.com/Microsoft/hcsshim/internal/ncproxy/store"
 	"github.com/Microsoft/hcsshim/internal/ncproxyttrpc"
 	"github.com/Microsoft/hcsshim/internal/oc"
 	"github.com/Microsoft/hcsshim/internal/uvm"
